src/main/model/AlgoChess/Tablero/Casillero.java
package model.AlgoChess.Tablero;

import model.AlgoChess.Equipos.Equipo;
import model.AlgoChess.Excepciones.CasilleroEnemigoExcepcion;
import model.AlgoChess.Excepciones.CasilleroOcupadoExcepcion;
import model.AlgoChess.Excepciones.CoordenadaFueraDeRangoExcepcion;
import model.AlgoChess.Unidades.Unidad;

public class Casillero{





    private boolean estaVacio = true;
    private int x;
    private int y;
    private Equipo equipo;
    private Tablero tablero;


    public Casillero(int xDado, int yDado, Equipo equipoDado, Tablero tableroDado){
        x = xDado;
        y = yDado;
        equipo = equipoDado;
        tablero = tableroDado;
    }
    public boolean estaLibre(){
        return estaVacio;
    }

    public boolean esDeEquipo(Equipo unEquipo){
        return equipo.esIgualA(unEquipo);
    }

    public void ocuparCasillero() throws CasilleroOcupadoExcepcion {

        if(!estaVacio){
            throw new CasilleroOcupadoExcepcion();
        }

        estaVacio = false;
    }

    public void colocarUnidad(Unidad unidad) throws CasilleroEnemigoExcepcion, CasilleroOcupadoExcepcion {
        if(!unidad.esDelEquipo(equipo)){
            throw new CasilleroEnemigoExcepcion();
        }

        ocuparCasillero();
    }

    public void vaciar(){
        estaVacio = true;
    }

    public int[] coordenadas(){
        return (new int[]{x, y});
    }

    public Casillero obtenerCasilleroSuperior() throws CoordenadaFueraDeRangoExcepcion {
            return tablero.conseguirCasillero(x, y+1);
    }

    public Casillero obtenerCasilleroInferior() throws CoordenadaFueraDeRangoExcepcion {
        return tablero.conseguirCasillero(x, y-1);
    }

    public Casillero obtenerCasilleroIzquierdo() throws CoordenadaFueraDeRangoExcepcion {
        return tablero.conseguirCasillero(x-1, y);
    }

    public Casillero obtenerCasilleroDerecho() throws CoordenadaFueraDeRangoExcepcion {
        return tablero.conseguirCasillero(x+1, y);
    }

}
